fix(gcp): make reportIfFoundPrivateKeyP12Format lazy load
- browser can not load node:fs
- node-forge is heavy module
